feat: 完成仪表盘前后端联调，使用真实数据库数据
## 后端改动
- 修复 RiskMetrics 实体类字段映射（@TableField注解）
- 完善 AssetServiceImpl 从数据库读取真实数据
  - getAssetOverview: 从 price_history 计算今日盈亏和年化收益
  - getCoreMetrics: 从 risk_metrics 表读取
  - getReturnCurve: 从 price_history 表读取
  - getRebalanceSuggestions: 根据偏离度生成调仓建议
- 新增 RebalanceSuggestionVO 数据结构
- 添加测试数据脚本 sql/dashboard_data.sql

## 前端改动
- 更新 axios 拦截器提取后端响应的 data 字段
- 更新所有仪表盘组件使用 store 中的真实数据
  - TotalAssetCard: 显示总资产和盈亏
  - PerformanceMetricsGrid: 累计收益/年化/偏离度
  - RiskMetricsCard: 风险指标
  - ModuleAllocationChart: 四象限配置
  - ModuleDetailsTable: 模块详情表格
  - AssetCurveChart: 收益曲线（支持周期切换）
  - RebalanceSuggestions: 动态平衡建议
- 修复 TypeScript 类型错误

+++ backend/asset-domain/src/main/java/com/asset/domain/vo/AssetOverviewVO.java
@@ -3,9 +3,11 @@ package com.asset.domain.vo;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 资产总览VO
+ * 遵循 backend/CLAUDE.md 规范
  */
 @Data
 public class AssetOverviewVO implements Serializable {
@@ -15,35 +17,35 @@ public class AssetOverviewVO implements Serializable {
     /**
      * 总资产
      */
-    private Double totalValue;
+    private BigDecimal totalValue;
 
     /**
      * 今日盈亏
      */
-    private Double todayProfit;
+    private BigDecimal todayProfit;
 
     /**
-     * 今日盈亏率
+     * 今日盈亏率（百分比数值，如 2.35 表示 2.35%）
      */
-    private Double todayProfitRate;
+    private BigDecimal todayProfitRate;
 
     /**
      * 累计盈亏
      */
-    private Double totalProfit;
+    private BigDecimal totalProfit;
 
     /**
-     * 累计盈亏率
+     * 累计盈亏率（百分比数值）
      */
-    private Double totalProfitRate;
+    private BigDecimal totalProfitRate;
 
     /**
-     * 年化收益率
+     * 年化收益率（百分比数值）
      */
-    private Double annualReturn;
+    private BigDecimal annualReturn;
 
     /**
-     * 基准对比
+     * 基准对比（百分比数值）
      */
-    private Double benchmarkComparison;
+    private BigDecimal benchmarkComparison;
 }
